+++ cryptpilot-core/src/provider/oidc.rs
@@ -106,8 +106,10 @@ impl KeyProvider for OidcKeyProvider {
 
     async fn get_key(&self) -> Result<Passphrase> {
         let cdh_bin_path = helper::find_cdh_binary_or_default();
-        #[cfg(not(test))]
-        if !std::path::Path::new(&cdh_bin_path).exists() {
+        if !cfg!(test)
+            && !std::path::Path::new(&cdh_bin_path).exists()
+            && std::env::var("CRYPTPILOT_TEST_MODE").is_err()
+        {
             anyhow::bail!(
                 "The confidential-data-hub binary not found, you may need to install it first."
             )
@@ -119,10 +121,12 @@ impl KeyProvider for OidcKeyProvider {
             .run()
             .await
             .context("failed to execute the command to get OIDC token");
-        #[cfg(not(test))]
-        let oidc_token = get_oidc_token_res?;
-        #[cfg(test)]
-        let oidc_token = { b"test_oidc_token".to_vec() };
+
+        let oidc_token = if cfg!(test) || std::env::var("CRYPTPILOT_TEST_MODE").is_ok() {
+            b"test_oidc_token".to_vec()
+        } else {
+            get_oidc_token_res?
+        };
 
         let oidc_token = String::from_utf8(oidc_token).context("failed to parse OIDC token")?;
 
@@ -160,11 +164,11 @@ impl KeyProvider for OidcKeyProvider {
             .await
             .context("failed to retrieve key using OIDC + KMS");
 
-        #[cfg(not(test))]
-        let key_u8 = get_secret_res?;
-
-        #[cfg(test)]
-        let key_u8 = { BASE64_STANDARD.encode(b"test").into_bytes() };
+        let key_u8 = if cfg!(test) || std::env::var("CRYPTPILOT_TEST_MODE").is_ok() {
+            BASE64_STANDARD.encode(b"test").into_bytes()
+        } else {
+            get_secret_res?
+        };
 
         let passphrase = (|| -> Result<_> {
             let key_base64 = String::from_utf8(key_u8)?;
